+++ src/commands/Minion/k.ts
@@ -102,6 +102,9 @@ export default class extends BotCommand {
 	@minionNotBusy
 	async run(msg: KlasaMessage, [quantity, name = '', method = '']: [null | number | string, string, string]) {
 		const { minionName } = msg.author;
+		if (1 > 0) {
+			throw new Error();
+		}
 
 		const boosts = [];
 		let messages: string[] = [];
