+++ src/lib/util/handleTripFinish.ts
@@ -9,16 +9,10 @@ import { handleGrowablePetGrowth } from '../growablePets';
 import { handlePassiveImplings } from '../implings';
 import clueTiers from '../minions/data/clueTiers';
 import { triggerRandomEvent } from '../randomEvents';
+import { runCommand } from '../settings/settings';
 import { ClientSettings } from '../settings/types/ClientSettings';
 import { ActivityTaskOptions } from '../types/minions';
-import {
-	channelIsSendable,
-	generateContinuationChar,
-	roll,
-	runCommand,
-	stringMatches,
-	updateGPTrackSetting
-} from '../util';
+import { channelIsSendable, generateContinuationChar, roll, stringMatches, updateGPTrackSetting } from '../util';
 import getUsersPerkTier from './getUsersPerkTier';
 import { sendToChannelID } from './webhook';
 import { activity_type_enum } from '.prisma/client';
